+++ src/Symfony/Component/Mailer/Bridge/Brevo/Transport/BrevoApiTransport.php
@@ -15,11 +15,14 @@ use Psr\EventDispatcher\EventDispatcherInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Mailer\Envelope;
 use Symfony\Component\Mailer\Exception\HttpTransportException;
+use Symfony\Component\Mailer\Exception\InvalidArgumentException;
 use Symfony\Component\Mailer\Header\MetadataHeader;
 use Symfony\Component\Mailer\Header\TagHeader;
 use Symfony\Component\Mailer\Header\TrackingHeader;
+use Symfony\Component\Mailer\RemoteTemplateEmail;
 use Symfony\Component\Mailer\SentMessage;
 use Symfony\Component\Mailer\Transport\AbstractApiTransport;
+use Symfony\Component\Mailer\Transport\RemoteTemplateTransportInterface;
 use Symfony\Component\Mime\Address;
 use Symfony\Component\Mime\Email;
 use Symfony\Component\Mime\Header\Headers;
@@ -31,7 +34,7 @@ use Symfony\Contracts\HttpClient\ResponseInterface;
 /**
  * @author Pierre TANGUY
  */
-final class BrevoApiTransport extends AbstractApiTransport
+final class BrevoApiTransport extends AbstractApiTransport implements RemoteTemplateTransportInterface
 {
     public function __construct(
         #[\SensitiveParameter] private string $key,
@@ -90,12 +93,24 @@ final class BrevoApiTransport extends AbstractApiTransport
     private function getPayload(Email $email, Envelope $envelope): array
     {
         $tracking = $this->getTracking($email->getHeaders());
+        $template = $email instanceof RemoteTemplateEmail ? $email->getRemoteTemplate() : null;
 
         $payload = [
             'sender' => $this->formatAddress($envelope->getSender()),
             'to' => $this->formatAddresses($this->getRecipients($email, $envelope), $tracking),
-            'subject' => $email->getSubject(),
         ];
+        if (null === $template || null !== $email->getSubject()) {
+            $payload['subject'] = $email->getSubject();
+        }
+        if (null !== $template) {
+            if (!ctype_digit($template->getReference())) {
+                throw new InvalidArgumentException(\sprintf('The Brevo API expects a numeric template id, "%s" given.', $template->getReference()));
+            }
+            $payload['templateId'] = (int) $template->getReference();
+            if ($template->getVariables()) {
+                $payload['params'] = $template->getVariables();
+            }
+        }
         if ($attachments = $this->prepareAttachments($email)) {
             $payload['attachment'] = $attachments;
         }
@@ -157,11 +172,13 @@ final class BrevoApiTransport extends AbstractApiTransport
                 continue;
             }
             if ('templateid' === $name) {
+                trigger_deprecation('symfony/brevo-mailer', '8.2', 'Using the "templateid" email header to select a Brevo template is deprecated, use a "%s" instead.', RemoteTemplateEmail::class);
                 $headersAndTags[$header->getName()] = (int) $header->getValue();
 
                 continue;
             }
             if ('params' === $name) {
+                trigger_deprecation('symfony/brevo-mailer', '8.2', 'Using the "params" email header to define the variables of a Brevo template is deprecated, use a "%s" instead.', RemoteTemplateEmail::class);
                 $headersAndTags[$header->getName()] = $header->getParameters();
 
                 continue;