+++ src/Symfony/Component/Mailer/Bridge/Resend/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+8.2
+---
+
+ * Add support for `RemoteTemplateEmail` to `ResendApiTransport`
+
 7.1
 ---